Simple whitespace tokenizer
I got inspired from reading
http://nitschinger.at/Text-Analysis-in-Rust-Tokenization.
The current simple whitespace tokenizer is basically that, but
more heavily based on adapting iterators. So we start with a
character iterator, wrap it in a character-data iterator (to
add byte and char offsets), and then wrap that in a token
iterator.

At the end of the day it's just a whitespace tokenizer, and
Rust already has library functions for splitting strings, so
it isn't particularly useful. Someday it will grow into a
full fledged rule-based tokenizer, though!